+++ web/app/components/header/facet-dropdown-list-item.hbs
@@ -1,0 +1,30 @@
+<li
+  role="option"
+  id="facet-dropdown-menu-item-{{this.id}}"
+  aria-selected={{this.isFocused}}
+  data-test-facet-dropdown-menu-item
+>
+  <LinkTo
+    tabindex="-1"
+    aria-checked={{@selected}}
+    @route={{this.currentRouteName}}
+    {{on "mouseover" this.onMouseover}}
+    {{! TODO: queryParam - pending a separate PR merge }}
+    class="facet-dropdown-menu-item-link {{if this.isFocused 'bg-red-500'}}"
+  >
+    <FlightIcon
+      @name="check"
+      class={{if @selected "visible" "invisible"}}
+    />
+    <div class="w-full flex justify-between overflow-hidden">
+      <span class="facet-dropdown-list-item-value">
+        {{@value}}
+      </span>
+      <Hds::BadgeCount
+        @text={{@count}}
+        class="opacity-75 ml-8 shrink-0"
+        @size="small"
+      />
+    </div>
+  </LinkTo>
+</li>
