+++ web/app/components/header/facet-dropdown.hbs
@@ -1,4 +1,10 @@
-<div class="relative" {{on "keydown" this.onKeydown}} ...attributes>
+<div
+  class="relative"
+  aria-owns="facet-dropdown-menu"
+  aria-haspopup="listbox"
+  aria-expanded={{this.dropdownIsShown}}
+  ...attributes
+>
   <Hds::Button
     data-test-facet-dropdown-trigger
     {{did-insert this.registerTrigger}}
@@ -10,66 +16,13 @@
     @text={{@label}}
   />
   {{#if this.dropdownIsShown}}
-    <div
+    <Header::FacetDropdownList
       {{dismissible
         dismissAction=this.hideDropdown
         related=this.triggerElement
       }}
-      {{did-insert this.registerPopover}}
-      class="facet-dropdown-popover
-        {{if this.inputIsShown 'large'}}
-        {{if (eq this.facetName 'status') 'medium'}}"
-    >
-      {{#if this.inputIsShown}}
-        <div class="border-b border-b-color-border-primary">
-          <Hds::Form::TextInput::Base
-            {{did-insert this.registerInput}}
-            {{on "input" (perform this.onInput)}}
-            {{on "focusin" this.resetMenuItemIndex}}
-            @value={{this.query}}
-            @type="search"
-            class="leading-none"
-            placeholder="Filter..."
-            class="facet-dropdown-input"
-          />
-        </div>
-      {{/if}}
-      <div class="overflow-auto h-full">
-        <ol
-          data-test-facet-dropdown-menu
-          role="menu"
-          aria-label={{@label}}
-          class="py-3 list-none"
-        >
-          {{#each-in this.shownFacets as |value attrs|}}
-            <li role="none" data-test-facet-dropdown-menu-item>
-              <LinkTo
-                role="menuitemradio"
-                tabindex="-1"
-                aria-checked={{attrs.selected}}
-                @route={{this.currentRouteName}}
-                {{! TODO: queryParam - pending a separate PR merge }}
-                class="facet-dropdown-menu-item-link"
-              >
-                <FlightIcon
-                  @name="check"
-                  class={{if attrs.selected "visible" "invisible"}}
-                />
-                <div class="w-full flex justify-between overflow-hidden">
-                  <span class="facet-dropdown-list-item-value">
-                    {{value}}
-                  </span>
-                  <Hds::BadgeCount
-                    @text={{attrs.count}}
-                    class="opacity-75 ml-8 shrink-0"
-                    @size="small"
-                  />
-                </div>
-              </LinkTo>
-            </li>
-          {{/each-in}}
-        </ol>
-      </div>
-    </div>
+      @label={{@label}}
+      @facets={{@facets}}
+    />
   {{/if}}
 </div>
